+++ PasswordReset.php
@@ -277,7 +277,7 @@ _printPageHead( _( 'Forgot your password?' ) );
 
 	<?php PopTable( 'header', _( 'Forgot your password?' ) ); ?>
 
-		<label><input type="text" name="email" id="email" size="25" maxlength="255" tabindex="1" pattern="[^ @]*@[^ @]*" required autofocus />
+		<label><input type="email" name="email" id="email" size="25" maxlength="255" tabindex="1" required autofocus />
 		<br />
 		<?php echo _( 'Email' ); ?></label>
 		<br />