+++ modules/Users/includes/Other_Info.inc.php
@@ -62,7 +62,7 @@ foreach ( (array) $fields_RET as $field )
 					issetVal( $value['EMAIL'] ),//issetVal( $value['CUSTOM_' . $field['ID']] ),
 					'staff[EMAIL]',
 					$field['TITLE'],
-					'size=22 maxlength=255 type="email" pattern="[^ @]*@[^ @]*" placeholder="' . AttrEscape( _( 'Email' ) ) . '"' .
+					'size=22 maxlength=255 type="email" placeholder="' . AttrEscape( _( 'Email' ) ) . '"' .
 						( ! empty( $_REQUEST['moodle_create_user'] ) || ! empty( $old_user_in_moodle )
 							|| $field['REQUIRED'] ? ' required' : '' ),
 					empty( $_REQUEST['moodle_create_user'] )